#339 rete tc node fix (self-loops)

+++ plugins/org.eclipse.viatra2.gtasm.patternmatcher.incremental.rete/src/org/eclipse/viatra2/gtasm/patternmatcher/incremental/rete/single/TransitiveClosureNode.java
@@ -88,7 +88,7 @@ public class TransitiveClosureNode extends SingleInputNode implements Clearable,
 				if (((IncSCCAlg<Object>) transitiveClosureAlgorithm).isIsolated(source)) {
 					graphDataSource.deleteNode(source);
 				}
-				if (((IncSCCAlg<Object>) transitiveClosureAlgorithm).isIsolated(target)) {
+				if (!source.equals(target) && ((IncSCCAlg<Object>) transitiveClosureAlgorithm).isIsolated(target)) {
 					graphDataSource.deleteNode(target);
 				}
 			}
